PREUNITA_NUEVA

+++ UanlSISM/UanlSISM/Models/P_OrdenCompra.cs
@@ -8,7 +8,7 @@ namespace UanlSISM.Models
     public partial class SISM_DET_REQUISICION
     {
         public short CANTIDAD_NUEVA { get; set; }
-        public float PREUNIT_NUEVA { get; set; }
+        public double PREUNIT_NUEVA { get; set; }
         public float TOTAL_NUEVA { get; set; }
         public bool CB_ELIMINAR { get; set; }
     }
@@ -16,7 +16,7 @@ namespace UanlSISM.Models
     public partial class SISM_DETALLE_OC
     {
         public short CANTIDAD_NUEVA { get; set; }
-        public float PREUNIT_NUEVA { get; set; }
+        public double PREUNIT_NUEVA { get; set; }
         public float TOTAL_NUEVA { get; set; }
         public bool CB_ELIMINAR { get; set; }
     }
